Refactor code structure with namespaces for modularity
This commit introduces namespaces to improve organization and structure within the project. The `IAgent` and `ISensor` interfaces have been moved to the `InvestigationGame.Interface` namespace, while the `BasicIranianAgent` and `BasicSensor` classes are now in `InvestigationGame.Agents` and `InvestigationGame.Sensors`, respectively.

The `InvestigationManager` class has been relocated to the `InvestigationGame.Manager` namespace and updated to reference the new interfaces. Additionally, the `Program.cs` file has been modified to include these namespaces, ensuring proper access to the necessary classes and interfaces.

These changes enhance code readability, maintainability, and overall organization.
